% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagilim.R
\name{analyze_distribution}
\alias{analyze_distribution}
\title{Veri setinde belirli bir sütunun dağılım analizini yapar ve görselleştirir.}
\usage{
analyze_distribution(file_path, column_name)
}
\arguments{
\item{file_path}{Excel dosyasının yolu}

\item{column_name}{Analiz yapılacak sütunun adı}
}
\description{
Veri setinde belirli bir sütunun dağılım analizini yapar ve görselleştirir.
}
\examples{
# Örnek kullanım:
# analyze_distribution("veri.xlsx", "Price")

}
