% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displayMissingDatas.R
\name{analyze_missing_data}
\alias{analyze_missing_data}
\title{Excel dosyasını okuma ve eksik veri analizini yapma}
\usage{
analyze_missing_data(file_path)
}
\arguments{
\item{file_path}{Excel dosyasının yolunu içeren bir karakter dizisi.}
}
\value{
Eksik veri analizini içeren bir veri çerçevesi.
}
\description{
Bu fonksiyon, belirtilen Excel dosyasını okur, her bir kolon için eksik
veri sayısını ve yüzdesini hesaplar, sonucu daha anlaşılır bir formatta
gösterir.
}
