% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaningInColumn.R
\name{clean_and_write_excel}
\alias{clean_and_write_excel}
\title{Excel dosyasını temizleyip yeni bir dosyaya yazdırma}
\usage{
clean_and_write_excel(input_file, output_file)
}
\arguments{
\item{input_file}{Giriş Excel dosyasının yolu ve adı.}

\item{output_file}{Çıktı olarak yazılacak temizlenmiş Excel dosyasının yolu ve adı.}
}
\value{
Temizlenmiş veri seti.
}
\description{
Bu fonksiyon, belirtilen Excel dosyasını okuyarak eksik verileri yönetir,
\%50'den fazla eksik veri içeren sütunları siler ve \%20'den fazla eksik
veri içeren sütunları çoğunluk değeri ile doldurarak temizlenmiş veriyi
yeni bir Excel dosyasına yazdırır.
}
\examples{
clean_and_write_excel("dosya_adi.xlsx", "temizlenmis_veri.xlsx")

}
\seealso{
\code{\link{read_excel}}, \code{\link{write_xlsx}}
}
\keyword{IO}
\keyword{data}
\keyword{file}
\keyword{manipulation}
