% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToNumeric.R
\name{convert_and_write_excel}
\alias{convert_and_write_excel}
\title{Belirli bir Excel dosyasındaki sütunu sayısal veri tipine dönüştürür ve Excel dosyasına geri yazar.}
\usage{
convert_and_write_excel(excel_file, column_name)
}
\arguments{
\item{excel_file}{Excel dosyasının yolu ve adı}

\item{column_name}{Dönüştürülecek sütunun adı}
}
\description{
Belirli bir Excel dosyasındaki sütunu sayısal veri tipine dönüştürür ve Excel dosyasına geri yazar.
}
\examples{
# Örnek kullanım:
# convert_and_write_excel("veri.xlsx", "Distance")

}
