% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete-column.R
\name{delete_column_from_excel}
\alias{delete_column_from_excel}
\title{Excel dosyasındaki belirtilen sütunu silen ve sonucu yeni bir Excel dosyasına yazan fonksiyon}
\usage{
delete_column_from_excel(input_file, column_name, output_file)
}
\arguments{
\item{input_file}{Giriş Excel dosyasının yolu ve adı.}

\item{column_name}{Silinecek sütunun adı.}

\item{output_file}{Çıktı olarak yazılacak Excel dosyasının yolu ve adı.}
}
\value{
None
}
\description{
Excel dosyasındaki belirtilen sütunu silen ve sonucu yeni bir Excel dosyasına yazan fonksiyon
}
\examples{
delete_column_from_excel("dosya_adi.xlsx", "sütun_adı", "güncellenmiş_veri.xlsx")

}
