% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filling-with-avarage.R
\name{fill_na_with_mean}
\alias{fill_na_with_mean}
\title{Belirtilen Excel dosyasındaki bir sütundaki eksik değerleri sütun ortalaması ile doldurarak yeni bir Excel dosyasına yazma}
\usage{
fill_na_with_mean(input_file, column_name, output_file)
}
\arguments{
\item{input_file}{Giriş Excel dosyasının yolu ve adı.}

\item{column_name}{Doldurulacak sütunun adı.}

\item{output_file}{Çıktı olarak yazılacak Excel dosyasının yolu ve adı.}
}
\value{
None
}
\description{
Belirtilen Excel dosyasındaki bir sütundaki eksik değerleri sütun ortalaması ile doldurarak yeni bir Excel dosyasına yazma
}
\examples{
fill_na_with_mean("dosya_adi.xlsx", "sütun_adı", "temizlenmis_veri.xlsx")

}
