% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filling-the-mode.R
\name{fill_na_with_mode}
\alias{fill_na_with_mode}
\title{Excel dosyasındaki belirtilen sütundaki eksik değerleri mod ile doldurarak yeni bir Excel dosyasına yazma}
\usage{
fill_na_with_mode(input_file, column_name, output_file)

fill_na_with_mode(input_file, column_name, output_file)
}
\arguments{
\item{input_file}{Giriş Excel dosyasının yolu ve adı.}

\item{column_name}{Doldurulacak sütunun adı.}

\item{output_file}{Çıktı olarak yazılacak Excel dosyasının yolu ve adı.}
}
\value{
None

None
}
\description{
Excel dosyasındaki belirtilen sütundaki eksik değerleri mod ile doldurarak yeni bir Excel dosyasına yazma

Excel dosyasındaki belirtilen sütundaki eksik değerleri mod ile doldurarak yeni bir Excel dosyasına yazma
}
\examples{
fill_na_with_mode("dosya_adi.xlsx", "sütun_adı", "temizlenmis_veri.xlsx")

fill_na_with_mode("dosya_adi.xlsx", "sütun_adı", "temizlenmis_veri.xlsx")

}
