% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karar-agaci-prediction.R
\name{karar_agaci_predict}
\alias{karar_agaci_predict}
\title{Predict Price using a trained decision tree model}
\usage{
karar_agaci_predict(model, new_data)
}
\arguments{
\item{model}{The trained decision tree model.}

\item{new_data}{A dataframe containing new data for prediction.}
}
\value{
A vector of predicted prices.
}
\description{
Predict Price using a trained decision tree model
}
\examples{
predictions <- predict(model, new_data)
}
