% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karar-agaci-train.R
\name{karar_agaci_train}
\alias{karar_agaci_train}
\title{Train a decision tree model to predict Price}
\usage{
karar_agaci_train(file_path)
}
\arguments{
\item{file_path}{The path to the Excel file containing the dataset.}
}
\value{
A trained decision tree model.
}
\description{
Train a decision tree model to predict Price
}
\examples{
model <- train("path_to_excel_file.xlsx")
}
