% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prediction.R
\name{predict_from_model}
\alias{predict_from_model}
\title{Eğitilen modeli kullanarak tahmin yapar.}
\usage{
predict_from_model(model, test_data)
}
\arguments{
\item{model}{Eğitilmiş model nesnesi.}

\item{test_data}{Test veri seti.}
}
\value{
Tahminler.
}
\description{
Eğitilen modeli kullanarak tahmin yapar.
}
