% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Train.R
\name{train_model}
\alias{train_model}
\title{Veri setini okur, işler ve model eğitir.}
\usage{
train_model(file_path)
}
\arguments{
\item{file_path}{Excel dosyasının yolu.}
}
\value{
Liste içinde model ve test veri seti.
}
\description{
Veri setini okur, işler ve model eğitir.
}
