% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{visualize_columns}
\alias{visualize_columns}
\title{Excel dosyasını ve iki sütun ismini parametre olarak alan ve bu iki sütuna göre görselleştirme yapan fonksiyon}
\usage{
visualize_columns(input_file, column_x, column_y)
}
\arguments{
\item{input_file}{Giriş Excel dosyasının yolu ve adı.}

\item{column_x}{X ekseni için kullanılacak sütunun adı.}

\item{column_y}{Y ekseni için kullanılacak sütunun adı.}
}
\value{
None
}
\description{
Excel dosyasını ve iki sütun ismini parametre olarak alan ve bu iki sütuna göre görselleştirme yapan fonksiyon
}
\examples{
visualize_columns("veri.xlsx", "sütun1", "sütun2")

}
